refactor: remove legacy policy api
