+++ cmd/infracost/diff_test.go
@@ -316,7 +316,7 @@ func TestDiffWithPolicyDataUpload(t *testing.T) {
 			IsJSON:      true,
 		}, func(ctx *config.RunContext) {
 			ctx.Config.PolicyV2APIEndpoint = ts.URL
-			ctx.Config.PoliciesV2Enabled = true
+			ctx.Config.PoliciesEnabled = true
 		},
 	)
 }
