+++ cmd/infracost/upload_test.go
@@ -212,7 +212,7 @@ func TestUploadWithBlockingTagPolicyFailure(t *testing.T) {
 		func(c *config.RunContext) {
 			c.Config.DashboardAPIEndpoint = dashboardApi.URL
 			c.Config.PolicyV2APIEndpoint = policyV2Api.URL
-			c.Config.PoliciesV2Enabled = true
+			c.Config.PoliciesEnabled = true
 		},
 	)
 }
@@ -245,7 +245,7 @@ func TestUploadWithTagPolicyWarning(t *testing.T) {
 		func(c *config.RunContext) {
 			c.Config.DashboardAPIEndpoint = dashboardApi.URL
 			c.Config.PolicyV2APIEndpoint = policyV2Api.URL
-			c.Config.PoliciesV2Enabled = true
+			c.Config.PoliciesEnabled = true
 		},
 	)
 }
@@ -278,7 +278,7 @@ func TestUploadWithBlockingFinOpsPolicyFailure(t *testing.T) {
 		func(c *config.RunContext) {
 			c.Config.DashboardAPIEndpoint = dashboardApi.URL
 			c.Config.PolicyV2APIEndpoint = policyV2Api.URL
-			c.Config.PoliciesV2Enabled = true
+			c.Config.PoliciesEnabled = true
 		},
 	)
 }
@@ -311,7 +311,7 @@ func TestUploadWithFinOpsPolicyWarning(t *testing.T) {
 		func(c *config.RunContext) {
 			c.Config.DashboardAPIEndpoint = dashboardApi.URL
 			c.Config.PolicyV2APIEndpoint = policyV2Api.URL
-			c.Config.PoliciesV2Enabled = true
+			c.Config.PoliciesEnabled = true
 		},
 	)
 }
