+++ internal/config/config.go
@@ -86,9 +86,8 @@ type Config struct {
 	DashboardEndpoint         string `yaml:"dashboard_endpoint,omitempty" envconfig:"DASHBOARD_ENDPOINT"`
 	UsageAPIEndpoint          string `yaml:"usage_api_endpoint,omitempty" envconfig:"USAGE_API_ENDPOINT"`
 	UsageActualCosts          bool   `yaml:"usage_actual_costs,omitempty" envconfig:"USAGE_ACTUAL_COSTS"`
-	PolicyAPIEndpoint         string `yaml:"policy_api_endpoint" envconfig:"POLICY_API_ENDPOINT"`
 	PolicyV2APIEndpoint       string `yaml:"policy_v2_api_endpoint,omitempty" envconfig:"POLICY_V2_API_ENDPOINT"`
-	PoliciesV2Enabled         bool
+	PoliciesEnabled           bool
 	TagPoliciesEnabled        bool
 	EnableDashboard           bool  `yaml:"enable_dashboard,omitempty" envconfig:"ENABLE_DASHBOARD"`
 	EnableCloud               *bool `yaml:"enable_cloud,omitempty" envconfig:"ENABLE_CLOUD"`
