+++ internal/providers/terraform/plan_json_provider.go
@@ -9,8 +9,6 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/infracost/infracost/internal/config"
-	"github.com/infracost/infracost/internal/prices"
-	"github.com/infracost/infracost/internal/scan"
 	"github.com/infracost/infracost/internal/schema"
 	"github.com/infracost/infracost/internal/ui"
 )
@@ -19,23 +17,17 @@ type PlanJSONProvider struct {
 	ctx                  *config.ProjectContext
 	Path                 string
 	includePastResources bool
-	scanner              *scan.TerraformPlanScanner
-	policyV2Client       *apiclient.PolicyV2APIClient
+	policyClient         *apiclient.PolicyAPIClient
 	logger               *logrus.Entry
 }
 
 func NewPlanJSONProvider(ctx *config.ProjectContext, includePastResources bool) *PlanJSONProvider {
-	var scanner *scan.TerraformPlanScanner
-	if ctx.RunContext.Config.PolicyAPIEndpoint != "" {
-		scanner = scan.NewTerraformPlanScanner(ctx.RunContext, ctx.Logger(), prices.GetPricesConcurrent)
-	}
-
-	var policyV2Client *apiclient.PolicyV2APIClient
+	var policyClient *apiclient.PolicyAPIClient
 	var err error
-	if ctx.RunContext.Config.PoliciesV2Enabled {
-		policyV2Client, err = apiclient.NewPolicyV2APIClient(ctx.RunContext)
+	if ctx.RunContext.Config.PoliciesEnabled {
+		policyClient, err = apiclient.NewPolicyAPIClient(ctx.RunContext)
 		if err != nil {
-			ctx.Logger().WithError(err).Errorf("failed to initialize policyV2 client")
+			ctx.Logger().WithError(err).Errorf("failed to initialize policy client")
 		}
 	}
 
@@ -43,8 +35,7 @@ func NewPlanJSONProvider(ctx *config.ProjectContext, includePastResources bool) 
 		ctx:                  ctx,
 		Path:                 ctx.ProjectConfig.Path,
 		includePastResources: includePastResources,
-		scanner:              scanner,
-		policyV2Client:       policyV2Client,
+		policyClient:         policyClient,
 		logger:               ctx.Logger(),
 	}
 }
@@ -108,16 +99,9 @@ func (p *PlanJSONProvider) LoadResourcesFromSrc(usage schema.UsageMap, j []byte,
 	project.PartialPastResources = partialPastResources
 	project.PartialResources = partialResources
 
-	if p.scanner != nil {
-		err := p.scanner.ScanPlan(project)
-		if err != nil {
-			p.logger.WithError(err).Debugf("Terraform project %s plan JSON scan failed", project.Name)
-		}
-	}
-
 	// use TagPolicyAPIEndpoint for Policy2 instead of creating a new config variable
-	if p.policyV2Client != nil {
-		err := p.policyV2Client.UploadPolicyData(project)
+	if p.policyClient != nil {
+		err := p.policyClient.UploadPolicyData(project)
 		if err != nil {
 			p.logger.WithError(err).Errorf("Terraform project %s failed to upload policy data", project.Name)
 		}
