--- internal/scan/testdata/simple_project/main.tf
@@ -1,23 +1,0 @@
-provider "aws" {
-  region                      = "us-east-1"
-  skip_credentials_validation = true
-  skip_requesting_account_id  = true
-  access_key                  = "mock_access_key"
-  secret_key                  = "mock_secret_key"
-}
-
-resource "aws_instance" "web_app" {
-  ami           = "ami-674cbc1e"
-  instance_type = "m5.4xlarge"
-
-  root_block_device {
-    volume_size = 50
-  }
-
-  ebs_block_device {
-    device_name = "my_data"
-    volume_type = "gp2"
-    volume_size = 1000
-    iops        = 800
-  }
-}
